Fix incorrect column mapping in KuduRecordCursor
